refactor(result)!: tighten ColumnType metadata widths and simplify DecimalValue
Narrow the column-type metadata to the widths Snowflake actually reports, and
drop the redundant per-value precision/scale carried by DecimalValue.

- ColumnType: precision and scale are now Option<u8> and length is Option<u32>
  (were Option<i64>). Snowflake reports precision <= 38, scale <= 37, and
  lengths up to 16 MiB, so these widths cover the real domains.
- Add a length to ColumnType::Binary; Snowflake reports a max byte length for
  it, mirroring Text.
- Mark the field-carrying variants (Fixed, Text, Time, TimestampNtz,
  TimestampLtz, TimestampTz, Binary) #[non_exhaustive] so future fields (such
  as the byte length) can be added without a breaking change.
- DecimalValue no longer carries precision/scale. Those are column metadata
  available on ColumnType::Fixed, not properties of a single value; DecimalValue
  is now the exact decimal string only.
- Deserialize the wire metadata (RawQueryResponseRowType) directly into the
  narrow widths, so an out-of-range value is a hard parse error rather than
  being silently degraded. These domains are stable Snowflake properties, unlike
  unrecognized type tags, which still fall back to ColumnType::Unknown.
- Fix the ColumnType::Text docs: length is the character count, not the byte
  length Snowflake also reports.
- Add result::test_data::column_type (bench-internals) so benches can build
  ColumnType values now that the variants are non_exhaustive.
- Replace the manual metadata probe with an assertive integration test,
  column_type_metadata_maps_declared_types.

BREAKING CHANGE: ColumnType precision and scale are now Option<u8> and length
is Option<u32>; ColumnType::Binary is a struct variant with a length field; the
field-carrying ColumnType variants are #[non_exhaustive]; and
DecimalValue::precision and DecimalValue::scale have been removed.

+++ src/statement/wire/response.rs
@@ -163,15 +163,15 @@ pub(crate) struct RawQueryResponseRowType {
     pub(crate) database: String,
     pub(crate) name: String,
     pub(crate) nullable: bool,
-    pub(crate) scale: Option<i64>,
+    pub(crate) scale: Option<u8>,
     #[allow(unused)]
     pub(crate) byte_length: Option<i64>,
-    pub(crate) length: Option<i64>,
+    pub(crate) length: Option<u32>,
     #[allow(unused)]
     pub(crate) schema: String,
     #[allow(unused)]
     pub(crate) table: String,
-    pub(crate) precision: Option<i64>,
+    pub(crate) precision: Option<u8>,
 
     #[serde(rename = "type")]
     pub(crate) data_type: String,
